+++ Changelog.md
@@ -21,6 +21,7 @@ Bugfixes:
  * SMTChecker: Fix false positive in contracts that cannot be deployed.
  * SMTChecker: Fix internal error on public getter returning dynamic data on older EVM versions where these are not available.
  * SMTChecker: Fix internal error on try-catch with function call in catch block.
+ * Control Flow Graph: Take internal calls to functions that always revert into account.
 
 
 AST Changes: